Adapted worflow for testing + some small changes
