+++ mango/tests/processing_module/test_object_functions.py
